feature: VocabularyImpl에 fake 데이터 추가
테스트용 데이터를 List<VocabularyImpl> 타입으로 추가하였다. 자체적으로 fake 데이터를 만들어 사용하던 테스트 코드를, 추가된 가짜 데이터를 사용하도록 수정하였다.
